+++ tests/fixtures/BrokenSyntax.pas
@@ -1,0 +1,10 @@
+unit BrokenSyntax;
+interface
+implementation
+
+procedure Broken;
+begin
+  := 42;
+end;
+
+end.